Made names of screenshot-files version neutral
The introduction uses a number of screenshots, full and partial,
to illustrate the RStudio interface and symbolize some webpages
of interest. Up to now, the names of these files contained
meta-data information on versions of RStudio, OS, CRAn etc.,
which makes updating a double pain.

I have changed the file names and their references in the .Rmd
master file to be version-neutral, and added instead a 00_README
in the figures-sub directory that lists all the relevant meta
information. This means that an annual/biannual/whatever update
can focus on the actual screenshots, which are fiddly enough
anyway, instead of painstakingly setting and updating file names.
(And no, given the update frequency and the lack of need for
concurrent versions, a simple meta-data file is more than enough).